add transforms sub-package, move SpectScaler into it
