+++ src/vak/__init__.py
@@ -20,6 +20,7 @@ from . import engine
 from . import io
 from . import metrics
 from . import models
+from . import transforms
 from . import util
 
 from .engine.model import Model
